pyproject.toml
[project]
name = "mitaibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.16"
dependencies = [
    "aiogram>=3.21.0",
    "aiogram-i18n>=1.4",
    "aiohttp>=3.12.15",
    "beanie>=2.0.0",
    "dependency>=0.0.3",
    "dependency-injector>=4.48.1",
    "fluent-compiler>=1.1",
    "fluent-runtime>=0.4.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-mongodb>=0.6.2",
    "langchain-openai>=0.3.30",
    "loguru>=0.7.3",
    "mcp[cli]>=1.13.0",
    "openai>=1.99.9",
    "pedalboard>=0.9.17",
    "pydantic-settings>=2.10.1",
    "pydub>=0.25.1",
    "pymongo>=4.14.0",
    "requests>=2.32.4",
    "soundfile>=0.13.1",
    "sox>=1.5.0",
]
